feat: [UEPR-187] [UEPR-192] [UEPR-194] Moved color and changed file names
