+++ src/_colors.scss
@@ -48,6 +48,7 @@ $ui-light-grayish-blue: hsla(215, 76%, 95%, 1);
 $ui-yellow: hsla(45, 100%, 50%, 1); // #FFBF00 Events Primary
 $ui-coral: hsla(350, 100%, 70%, 1); // #FF6680 More Blocks Primary
 $ui-coral-dark: hsla(350, 100%, 60%, 1); // #FF3355 More Blocks tertiary
+$ui-coral-warm: hsla(360, 80%, 65%, 1); // #ED5E5E
 
 $ui-white: hsla(0, 100%, 100%, 1); //#FFF
 $ui-white-15percent: hsla(0, 100%, 100%, .15); //#FFF
@@ -58,6 +59,7 @@ $ui-border: hsla(0, 0%, 85%, 1); //#D9D9D9
 $ui-border-light-blue: hsla(215, 50%, 90%, 1); //#D9E3F2
 
 /* modals */
+$ui-mint-green-darker: hsl(163, 85%, 35%);
 $ui-mint-green: hsl(163, 69%, 44%);
 $ui-light-mint: hsl(163, 53%, 67%);
