+++ src/views/splash/presentation.jsx
@@ -37,7 +37,7 @@ const HOCTopBanner = require('./hoc/top-banner.jsx');
 const HOCMiddleBanner = require('./hoc/middle-banner.jsx');
 const FeaturesBanner = require('./features/features-banner.jsx');
 
-const TermsOfUseBanner = require('../../components/modal/terms-of-use/terms-of-use-banner.jsx').default;
+const TermsOfUseBanner = require('../../components/modal/terms-of-use/banner.jsx').default;
 
 require('./splash.scss');
